+++ src/main/java/hexlet/code/app/repository/LabelRepository.java
@@ -1,9 +1,12 @@
 package hexlet.code.app.repository;
 
+
 import hexlet.code.app.model.Label;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LabelRepository extends JpaRepository<Label, Integer> {
+public interface LabelRepository extends JpaRepository<Label, Long> {
+
 }
+
