+++ src/main/java/hexlet/code/app/repository/TaskStatusRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaskStatusRepository extends JpaRepository<TaskStatus, Integer> {
+public interface TaskStatusRepository extends JpaRepository<TaskStatus, Long> {
 }
